Refactor auth service import paths and optimize components
Moved `authService` to `api` folder for better structure and updated imports across components. Adjusted `BreakDownLPN` dependency array and improved column widths in `LPN` table for responsive design. These changes enhance code organization, performance, and user experience.
